Add files via upload
This is a fun little project about zoos. In this project there are 3 zoos, Denver Zoo, San Fran Zoo, and Atlanta Zoo. Each zoo has an array of animals that they have and they also have an array of animals that they want. The program will ask the user what baby they want to be born and from what zoo, then it will check which zoo has the animal in their want array and move the animal born to that zoo's have array.