Animals assignment
This assignment was made to show my experience and learning in inheritance, abstract and virtual method, and enums.

The instructions for this assignment. 
Create an animal class. Create a dog and a cat class that demonstrates inheritance. Animal class must implement name, height, weight, color. Implement bark volume as an enum in the dog class. Implement fur length as an enum in the cat class. Create a specific method in the dog and cat class to determine amount of food eaten per day. A cat will eat 1/4 oz of food for each pound they weigh. A dog will each 1 oz of food for each pound they weight. The method should be defined (made/created) in the animal class but implemented (has code inside to do work) in the cat and dog class.